Update README.md
I am adding a link to the colab for Social Impacts of Artificial Intelligence https://colab.research.google.com/drive/1bSo9oXpB7fHjPB5UZGKJAcyA0zXHGjZO?usp=sharing. I do not have rights to add the file directly. Please accept this PR or add the colab as a file to this repo and use the same format as for the other practicals.